fix(auth): drop redundant toLong() on Long constants breaking -Werror build
RESET_VALID_HOURS and INVITE_VALID_DAYS are already Long; .toLong()
triggered redundant-conversion warnings that fail compileKotlin under
warningsAsErrors. Introduced by 21156ebf's earlier gate repair.

Validated: :backend:compileKotlin -PwarningsAsErrors=true

Closes #362, ref #362

+++ backend/src/main/kotlin/com/companyb/companyapp/service/AuthService.kt
@@ -186,7 +186,7 @@ object AuthService {
                     // Reset-token validity is credential lifecycle (#322): auth owns JVM-clock reads.
                     expiresAt =
                         OffsetDateTime.ofInstant(
-                            Instant.now().plus(RESET_VALID_HOURS.toLong(), ChronoUnit.HOURS),
+                            Instant.now().plus(RESET_VALID_HOURS, ChronoUnit.HOURS),
                             ZoneOffset.UTC,
                         ),
                     createdBy = null,
